+++ ToDoList_F/src/main/java/com/todo/service/TeamTodoService.java
@@ -10,10 +10,10 @@ public interface TeamTodoService {
 	TodoEntity2 insertTodo(TodoEntity2 todo);
 	// r
 	//TodoEntity2 showTodoAll();
-	List<TodoEntity2> showTodoById(int team_id); 
+	List<TodoEntity2> showTodoById(Long team_id); 
 	// u
-	void updateTodo(int team_id, int content_id, TodoEntity2 todo);
-	void updateCompleted(int team_id, int content_id);
+	void updateTodo(Long team_id, Long content_id, TodoEntity2 todo);
+	void updateCompleted(Long team_id, Long content_id);
 	// d
-	void deleteTodo(int team_id, int content_id);
+	void deleteTodo(Long team_id, Long content_id);
 }
